Create IDEA
here is my idea of what rules that I have instigated into this game . The most interestimg part is that there ae no winners for this game 😂.
I have also included negative decrement of positions randomly when an individual rolled the dice.
The final snake of this game lies at final peice of the board.
The use can set their own personal board with nXn size.
It is not a single player game . Multiple players can play the simultaneously.
Instead of the boring automatic code ouput which results the winner without any user involvement. I had the idea to make it user responsive. 
Suppose their three players in the game, each have their own turn to roll the dice  and after every turn they have the output displayed whateve the case(snake bitten or moved up the ladder o changed the position)
As there is no clearr winner  in this game, the idea to entertainment is this torture that there is no winner of this game.